requirements.txt
langgraph==0.2.50
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
langchain-core==0.3.15
gradio==4.44.0
python-dotenv==1.0.0
pydantic==2.9.2
typing-extensions==4.12.2
httpx==0.27.0
asyncio-mqtt==0.16.2
playwright==1.48.0
requests==2.32.3
beautifulsoup4==4.12.2
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.2
Pillow==10.4.0
aiofiles==24.1.0
nest-asyncio==1.6.0
